feat: structures
- Updates structures of main app.
- Update path for integration test.
